+++ javascript/src/index.ts
@@ -1,4 +1,5 @@
 import Argument from './Argument.js'
+import Compiler from './Compiler.js'
 import CucumberExpression from './CucumberExpression.js'
 import CucumberExpressionGenerator from './CucumberExpressionGenerator.js'
 import ExpressionFactory from './ExpressionFactory.js'
@@ -11,6 +12,7 @@ import { Expression } from './types.js'
 
 export {
   Argument,
+  Compiler,
   CucumberExpression,
   CucumberExpressionGenerator,
   Expression,
